refactor(shared): extract coerce_user_valves into shared.valves
User valve payloads can arrive as model instances, compatible BaseModel instances, or raw dictionaries. Share the coercion helper used by agent plugins.
