+++ src/owui_ext/tools/sub_agent.py
@@ -60,6 +60,7 @@ from owui_ext.shared.tool_servers import (
     normalize_direct_tool_servers,
     resolve_direct_tool_servers_from_request_and_metadata,
 )
+from owui_ext.shared.valves import coerce_user_valves
 
 log = logging.getLogger(__name__)
 
@@ -78,21 +79,6 @@ class SubAgentTaskItem(BaseModel):
 # ============================================================================
 # Helper functions (outside class - AI cannot invoke these)
 # ============================================================================
-
-
-def coerce_user_valves(raw_valves: Any, valves_cls: Type[BaseModel]) -> BaseModel:
-    """Normalize raw user valves into the target valves class."""
-    if isinstance(raw_valves, valves_cls):
-        return raw_valves
-    if isinstance(raw_valves, BaseModel):
-        try:
-            data = raw_valves.model_dump()
-        except Exception:
-            data = {}
-        return valves_cls.model_validate(data)
-    if isinstance(raw_valves, dict):
-        return valves_cls.model_validate(raw_valves)
-    return valves_cls.model_validate({})
 
 
 def normalize_parallel_sub_agent_tasks(tasks: Any) -> tuple[Optional[list[dict[str, str]]], Optional[str]]:
