+++ tools/sub_agent.py
@@ -540,6 +540,30 @@ def build_direct_tools_dict(
         _tool_servers_log.info("No direct tools loaded from tool_servers")
     return direct_tools
 
+# --- inlined from src/owui_ext/shared/valves.py (owui_ext.shared.valves) ---
+from typing import Any, Type
+from pydantic import BaseModel
+def coerce_user_valves(raw_valves: Any, valves_cls: Type[BaseModel]) -> BaseModel:
+    """Normalize raw user valves into the target valves class.
+
+    Open WebUI hands ``raw_valves`` over from filter context, where it can
+    arrive as the target class itself, a different ``BaseModel`` subclass
+    (when the user-valve schema has drifted between plugin versions), a raw
+    dict, or anything else. Always return a fresh ``valves_cls`` instance so
+    callers can rely on the field set being current.
+    """
+    if isinstance(raw_valves, valves_cls):
+        return raw_valves
+    if isinstance(raw_valves, BaseModel):
+        try:
+            data = raw_valves.model_dump()
+        except Exception:
+            data = {}
+        return valves_cls.model_validate(data)
+    if isinstance(raw_valves, dict):
+        return valves_cls.model_validate(raw_valves)
+    return valves_cls.model_validate({})
+
 log = logging.getLogger(__name__)
 
 
@@ -557,21 +581,6 @@ class SubAgentTaskItem(BaseModel):
 # ============================================================================
 # Helper functions (outside class - AI cannot invoke these)
 # ============================================================================
-
-
-def coerce_user_valves(raw_valves: Any, valves_cls: Type[BaseModel]) -> BaseModel:
-    """Normalize raw user valves into the target valves class."""
-    if isinstance(raw_valves, valves_cls):
-        return raw_valves
-    if isinstance(raw_valves, BaseModel):
-        try:
-            data = raw_valves.model_dump()
-        except Exception:
-            data = {}
-        return valves_cls.model_validate(data)
-    if isinstance(raw_valves, dict):
-        return valves_cls.model_validate(raw_valves)
-    return valves_cls.model_validate({})
 
 
 def normalize_parallel_sub_agent_tasks(tasks: Any) -> tuple[Optional[list[dict[str, str]]], Optional[str]]:
